Replace Map<Character> by CharObjectHashMap and Set<Character> by CharHashSet. (#13420)
Also optimize the character replacement in JapaneseKatakanaUppercaseFilter.

+++ lucene/core/src/test/org/apache/lucene/util/hppc/TestLongObjectHashMap.java
@@ -335,7 +335,7 @@ public class TestLongObjectHashMap extends LuceneTestCase {
   /* */
   @Test
   public void testEmptyKey() {
-    final int empty = 0;
+    final long empty = 0;
 
     map.put(empty, value1);
     assertEquals(1, map.size());
